Major modifications and restructuring.
Added additional elements to gitignore.
Added additional configuration item to enable/disable crawling.
Defaulted to log to SQL.
Modified the database to support emojis on MariaDB
Added support to add channels the account is already a member of to the
monitoring list.
All channels are now always being monitored
In the event the channel posts a message to a group, it will be logged
as a generic user account.